+++ CHANGELOG.md
@@ -18,6 +18,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 ### Changed
 
 - `RDF.Turtle.Decoder` saves the base IRI now
+- `RDF.Turtle.Encoder` now takes the base IRI to be serialized from the graph when 
+  no base IRI is given with the `base` option 
 
 
 [Compare v0.6.1...HEAD](https://github.com/marcelotto/rdf-ex/compare/v0.6.1...HEAD)